requirements.txt
# Core Agent Development Kit
google-genai>=1.0.0
google-adk>=0.1.0

# Environment management
python-dotenv>=1.0.0

# Testing and evaluation
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality (optional but recommended)
black>=23.0.0
pylint>=2.17.0
mypy>=1.4.0

# Additional utilities
typing-extensions>=4.7.0